requirements.txt
streamlit>=1.28.0
pypdf>=3.17.0
Pillow>=10.0.0
ghostscript>=0.7
